requirements.txt
Django==4.2.6
djangorestframework==3.14.0 
django-cors-headers==4.3.1
django_filter==23.5
djangorestframework-simplejwt==5.3.1
djoser